Refactor CartItemDto to use string type for ItemId
The code changes in the `CartItemDto.cs` file modify the `ItemId` property type from `long` to `string`. This change is made to align the data type of the `ItemId` property with the actual data being used.

This commit message provides a concise summary of the changes made, starting with an action verb (`Refactor`) to indicate the type of change. It also includes a brief explanation of the purpose behind the changes.